+++ src/utils/constants.js
@@ -3,17 +3,17 @@ export const BING_API = "ArNqsYDx-rtxpMbR4ddz8SyY4-dv8-JK35KErFW3GIU7_UwgaCVz8Bj
 export const PRAYERS_ARR = ['Fajr', 'Dhuhr', 'Asr', 'Maghrib', 'Isha'];
 export const REPORT_FB = "Report Crash";
 export const P_MENUS = [
-    {label : 'Home', page : 'Home'}, 
-    {label : 'Finetune', page : 'setFTmodal'},
-    {label : 'Traveller Onboard', page : 'Travel'},
-    {label : 'Subscribe', page: 'setmodal'},    
+    {id: 0, label : 'Home', page : 'Home', description: 'Home page'}, 
+    {id: 1, label : 'Traveller Onboard', page : 'Travel', description: 'Traveller Onboard page'},
+    {id: 2, label : 'Subscribe', page: 'setmodal', description: 'Subscribe page'},    
     // {label : 'Fine Tune', page : 'Fine Tune'}, 
     // {label : 'Travel Times', page : 'Travel'}, 
     // {label : 'Nearest Mosque', page : 'Nearest Mosque'}, 
     // {label : 'Map', page : 'Map'} 
 ];
 export const P_MENUS_SEC = [
-    {label : 'Messy Lab', page: 'Lab'},
-    {label : 'Start Over', page: 'callfunc'}  
+    {id: 0, label : 'Finetune', page : 'setFTmodal', description: 'Finetune page'},
+    {id: 1, label : 'Messy Lab', page: 'Lab', description: 'lab page'},
+    {id: 2, label : 'Start Over', page: 'callfunc', description: 'Startover  page'}  
 ];
 export const FT_PRAYER = "api_prayer-results";